fixing PR comments
