+++ express/blocks/hero-color/hero-color.js
@@ -10,35 +10,8 @@
  * governing permissions and limitations under the License.
  */
 
-import {
-  createTag,
-} from '../../scripts/scripts.js';
-
-function isDarkOverlayReadable(colorString) {
-  let r;
-  let g;
-  let b;
-
-  if (colorString.match(/^rgb/)) {
-    const colorValues = colorString.match(
-      /^rgba?\((\d+),\s*(\d+),\s*(\d+)(?:,\s*(\d+(?:\.\d+)?))?\)$/,
-    );
-    [r, g, b] = colorValues.slice(1);
-  } else {
-    const hexToRgb = +`0x${colorString
-      .slice(1)
-      .replace(colorString.length < 5 ? /./g : '', '$&$&')}`;
-    // eslint-disable-next-line no-bitwise
-    r = (hexToRgb >> 16) & 255;
-    // eslint-disable-next-line no-bitwise
-    g = (hexToRgb >> 8) & 255;
-    // eslint-disable-next-line no-bitwise
-    b = hexToRgb & 255;
-  }
-
-  const hsp = Math.sqrt(0.299 * (r * r) + 0.587 * (g * g) + 0.114 * (b * b));
-  return hsp > 140;
-}
+import { createTag } from '../../scripts/scripts.js';
+import isDarkOverlayReadable from '../../scripts/color-tools.js';
 
 function changeTextColorAccordingToBg(
   primaryColor,
